refactor: moving some funcs in execution.py to controller_utils
